Prevent meta-class error in top-level imports.
Extended the `ImportError` handling hack to correctly handle the
case where we cannot import submodules AND someone is extending
one of the exported classes.  Previously ContentMixin and
ContentSettings were functions when ImportError was raised which
causes really confusing errors.

The correct solution here is to stop exporting stuff at the top-
level and require clients to import from modules but that is for
another day.