Remove usage of annotations future.
Though we weren't using the affected feature, I am going to avoid
``from __future__ import annotations`` since it is likely going to be
replaced with a different approach.  The only two things that we were
getting out of it is using `dict[]` instead of `typing.Dict[]` and using
`typing.Literal` in Python 3.7.  Neither is compelling IMO.

https://discuss.python.org/t/type-annotations-pep-649-and-pep-563/